Split binary and unary booleans into separate files
Too confusing and they had nothing in common as it turned out
