+++ lib/mlibrary_search_parser/node.rb
@@ -3,7 +3,8 @@ require_relative 'node/base'
 require_relative 'node/tokens'
 require_relative 'node/unparseable'
 require_relative 'node/empty'
-require_relative 'node/boolean'
+require_relative 'node/binary'
+require_relative 'node/unary'
 require_relative 'node/fielded'
 require_relative 'node/search'
